pyproject.toml
[project]
name = "article"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.16",
    "aioredis>=2.0.1",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "dnspython>=2.7.0",
    "elasticsearch==7.17.12",
    "email-validator>=2.2.0",
    "extra-streamlit-components>=0.1.71",
    "fastapi>=0.115.12",
    "greenlet>=3.1.1",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "hiredis>=3.1.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "loguru>=0.7.3",
    "mongoengine>=0.29.1",
    "motor>=3.7.0",
    "mypy>=1.15.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.11.3",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.40",
    "starlette-admin>=0.14.1",
    "strawberry-graphql>=0.263.2",
    "streamlit>=1.44.1",
    "typer>=0.15.2",
    "uvicorn>=0.34.0",
]
